[cuttlefish] test : improved level loading
