+++ src/cuttlefish/test/src/game/level1.hh
@@ -1,0 +1,13 @@
+#ifndef __level1__
+#define __level1__
+
+namespace cuttlefish
+{
+    class Window;
+
+    void runLevel1(Window &window_p);
+} // namespace cuttlefish
+
+
+
+#endif
